Decouple some utilities
