+++ src/utils/utils.tsx
@@ -1,25 +1,12 @@
-import { turf } from './index'
-import hmaPoly from './../HMA.json'
-import { aqiLabels, walkSpeed } from './../constants'
+import { walkSpeed } from './../constants'
 import { MapMouseEvent, Map, PointLike } from 'mapbox-gl'
 
-export const getNoiseIndexLabel = (ni: number): string | undefined => {
-  if (ni < 0.15) return 'very quiet'
-  if (ni < 0.3) return 'quiet'
-  if (ni < 0.5) return 'moderate noise'
-  if (ni < 0.65) return 'high noise'
-  if (ni < 0.75) return 'very high noise'
-  if (ni >= 0.75) return 'extreme noise'
-}
-
-export const getAqiLabel = (aqi: number): string => {
-  if (aqi <= 0) return ''
-  if (aqi <= 2.0) return aqiLabels[1]
-  if (aqi <= 3.0) return aqiLabels[2]
-  if (aqi <= 4.0) return aqiLabels[3]
-  if (aqi <= 5.0) return aqiLabels[4]
-  if (aqi > 5.0) return aqiLabels[5]
-  return ''
+const concatSign = (number: number): string => {
+  if (number < 0) {
+    return '-' + String(number)
+  } else if (number > 0) {
+    return '+' + String(number)
+  } else return String(number)
 }
 
 export const getDurationStringFromDist = (m: number, showSeconds: boolean = false, withSign: boolean = false): string => {
@@ -38,60 +25,6 @@ export const getDurationStringFromDist = (m: number, showSeconds: boolean = fals
     formattedDuration = withSign === true ? concatSign(roundedMins) : String(roundedMins)
   }
   return formattedDuration + ' ' + unit
-}
-
-const concatSign = (number: number): string => {
-  if (number < 0) {
-    return '-' + String(number)
-  } else if (number > 0) {
-    return '+' + String(number)
-  } else return String(number)
-}
-
-const roundTo = (number: number, digits: number): number => {
-  return Math.round(number * (10 * digits)) / (10 * digits)
-}
-
-export const getFormattedDistanceString = (m: number, withSign: boolean = false): string => {
-  let distance
-  let unit
-  if (Math.abs(m) >= 950) {
-    const km = m / 1000
-    distance = roundTo(km, 1)
-    unit = ' km'
-  } else if (Math.abs(m) > 60) {
-    distance = Math.round(m / 10) * 10
-    unit = ' m'
-  } else {
-    distance = Math.round(m)
-    unit = ' m'
-  }
-  const distanceString = withSign === true ? concatSign(distance) : String(distance)
-  return distanceString + unit
-}
-
-export const getFormattedAqiExpDiffRatio = (aqc_diff_rat: number): string => {
-  if (Math.round(aqc_diff_rat) === 0) {
-    return '-' + String(Math.round(aqc_diff_rat))
-  } else {
-    return String(Math.round(aqc_diff_rat))
-  }
-}
-
-export const getOriginCoordsFromFC = (FC: PointFeatureCollection): [number, number] | null => {
-  const origin = FC.features.filter(feat => feat.properties.type === 'orig')
-  if (origin.length === 0) return null
-  const coords = origin[0].geometry.coordinates
-  // @ts-ignore
-  return coords.map(coord => Math.round(coord * 100000) / 100000)
-}
-
-export const getDestCoordsFromFC = (FC: PointFeatureCollection): [number, number] | null => {
-  const dest = FC.features.filter(feat => feat.properties.type === 'dest')
-  if (dest.length === 0) return null
-  const coords = dest[0].geometry.coordinates
-  // @ts-ignore
-  return coords.map(coord => Math.round(coord * 100000) / 100000)
 }
 
 export const getLayersFeaturesAroundClickE = (layers: string[], e: MapMouseEvent, tolerance: number, map: Map) => {
@@ -146,18 +79,4 @@ export const getInitialLengthLimit = (lengthLimits: LengthLimit[], pathCount: nu
     }
   }
   return lengthLimits[lengthLimits.length - 1]
-}
-
-export const origDestWithinSupportedArea = (origDestFC: PointFeatureCollection) => {
-  const origin = origDestFC.features.filter(feat => feat.properties.type === 'orig')
-  const dest = origDestFC.features.filter(feat => feat.properties.type === 'dest')
-  // @ts-ignore
-  const extentFeat: PolygonFeature = hmaPoly.features[0]
-  if (origin.length > 0 && !turf.within(origin[0], extentFeat)) {
-    return 'Origin is outside the supported area'
-  }
-  if (dest.length > 0 && !turf.within(dest[0], extentFeat)) {
-    return 'Destination is outside the supported area'
-  }
-  return null
 }
